# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tether API",
    "description": "\n# tether API\n\ntether helps you hold yourself accountable to keep your phone away from your bedroom at night.\n\n## Overview\n\nThis API runs on a Raspberry Pi and provides:\n\n1. **Proximity Detection**: Check if your phone is near the Raspberry Pi via Bluetooth\n2. **Emergency Passes**: A limited number of monthly passes for legitimate exceptions\n3. **Configuration**: Manage Bluetooth devices and settings\n\n## For AI Agents (MCP)\n\nIf you're accessing this API via MCP tools:\n\n- **checkProximity**: Verify the phone is in its designated spot. Returns `is_nearby: true` when close.\n- **getPasses**: Check how many emergency passes remain this month.\n- **usePass**: Use when user has legitimate reason (on-call, emergency). Requires a reason.\n- **getPassHistory**: Review past pass usage to identify patterns.\n\n## Design Philosophy\n\n- **Lazy evaluation**: Bluetooth checks only happen when requested\n- **Intentional friction**: Passes require reasons to encourage mindfulness\n- **Delayed effects**: Pass count changes only apply next month to prevent gaming\n",
    "contact": {
      "name": "Jeffrey",
      "email": "jeffrey@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local tether server"
    }
  ],
  "paths": {
...
    "/system/restart": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Request system restart",
        "description": "Initiates a system restart. The restart can be delayed by up to 60 seconds. This is useful for applying configuration changes that require a restart.",
        "operationId": "restartSystem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Restart accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid delay value"
          },
          "503": {
            "description": "Restart not available"
          }
        }
      }
    },
...
    "/system/ticket": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get dumbpipe ticket",
        "description": "Returns the dumbpipe ticket for establishing remote P2P connections via iroh. This ticket is used by the MCP server to connect to this tether instance.",
        "operationId": "getDumbpipeTicket",
        "responses": {
          "200": {
            "description": "Ticket retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DumbpipeTicketResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dumbpipe not available"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DumbpipeTicketResponse": {
        "type": "object",
        "description": "Dumbpipe ticket response.",
        "required": [
          "available"
        ],
        "properties": {
          "available": {
            "type": "boolean",
            "description": "Whether dumbpipe is available.",
            "example": true
          },
          "expires_at_utc": {
            "type": [
              "string",
              "null"
            ],
            "description": "When the ticket expires (if applicable).",
            "example": "2025-01-15T04:30:00Z"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Message if not available."
          },
          "node_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The node ID of this tether instance.",
            "example": "n0abc123..."
          },
          "ticket": {
            "type": [
              "string",
              "null"
            ],
            "description": "The dumbpipe ticket for remote access.\nThis is a base32-encoded iroh ticket.",
            "example": "blobfd23abc..."
          }
        },
        "example": {
          "expires_at_utc": "2025-01-15T04:30:00Z",
          "node_id": "n0abc123...",
          "ticket": "blobfd23abc..."
        }
      },
...
      "RestartRequest": {
        "type": "object",
        "description": "System restart request.",
        "properties": {
          "delay_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Delay before restart in seconds (0-60).",
            "example": 5,
            "maximum": 60,
            "minimum": 0
          }
        },
        "example": {
          "delay_secs": 5
        }
      },
      "RestartResponse": {
        "type": "object",
        "description": "System restart response.",
        "required": [
          "accepted",
          "message",
          "delay_secs"
        ],
        "properties": {
          "accepted": {
            "type": "boolean",
            "description": "Whether the restart request was accepted.",
            "example": true
          },
          "delay_secs": {
            "type": "integer",
            "format": "int32",
            "description": "Delay before restart.",
            "example": 5,
            "minimum": 0
          },
          "message": {
            "type": "string",
            "description": "Message about the restart.",
            "example": "System will restart in 5 seconds"
          }
        },
        "example": {
          "accepted": true,
          "delay_secs": 5,
          "message": "System will restart in 5 seconds"
        }
      },
...
    }
  },
...
}